refactor(cli): align parser filenames
